Langchain/requirements.txt
langchain
langchain_community
streamlit
ollama
python-dotenv